Updated openstack/openstack Project: openstack-infra/shade  9db26ef1e2c7a33ad01eaf2a9cd27ff1ed56be97
Fix intermittent error in unit tests

After the introduction of unit tests for the _neutron_exceptions
context, unit tests fail randomly with the following error:

    AttributeError: <class 'shade.OpenStackCloud'> does not have the
    attribute 'mock_me'

This error is due to the wrong location of test__neutron_excetpions
tests (i.e. the TestShadeOperator class rather than the TestShade
class).
Moreover, as pointed out on IRC we can get rid of "fake" methods to test
 the  _neutron_exception context, just using the existing code.

Change-Id: I2cfdcfce606272fbdbdf8a463db895d3d5a4a1e3