Updated openstack/openstack Project: openstack-infra/subunit2sql  1875f963b2cc3ebcd9fce5e25f2912894fbe5fc7
Add graph command to graph daily test count over time

This commit adds a new graph command to generate a line graph showing
the number of tests executed daily over time. As part of this it was
necessary to add a new DB api method to return the daily run count
time series to be used for graphing.

Change-Id: I886a9bf1558aa0cf7e1176519ab8a79bf62a2e23