Updated openstack/openstack Project: openstack/ironic-python-agent  4c1d2cdb190a894e7cc83ed5d30061a53fe9e09b
IPA:'shred' utility to use configured iterations

Today, there is no option to configure number of iterations to be
done for shred block device erasing and defaults it to 1. This patch
adds a configuration option to change the number of passes to be done
to erase a block device.

Change-Id: I1921d33a6b364c4682b6c9baaf61ac092cfa11d7
Partial-Bug:#1465130