Updated openstack/openstack Project: openstack/nova  82d4e9026766970e670e32142078788a72437f56
VMware: Remove configuration check

Checking if use_linked_clone is None doesn't make any sense because this
is boolean config property which is either True or False (default is
True). If the user hasn't specified a boolean value, the configuration
framework will raise ConfigFileValueError, so there is no need to make
this check in Nova.

Change-Id: I9af15e16cc4ea2779c74b38ad40edf1a2d06d1d1