Updated openstack/openstack Project: openstack/ironic  25933c608f2b20e48c185408f6d91bf05ae14ed1
Add disk layout check on re-provisioning

The preserve ephemeral logic does not check whether the
disk layout was changed since the last time the node was
provisioned. Additionally we have introduced a 
driver_internal_info['instance'] object to store instance's
disk configration and which is purged when tearing down an
instance.

Change-Id: Iecdb8db37b22c37275be3197950bbf899aa62a53
Closes-Bug: #1327260