Updated openstack/openstack Project: openstack/nova  beabe5dd276b6088bf36cd9991751c158f106a88
network: add more debug logging context for race bug 1249065

The logging added in commit c13f42e36cddd820b475ffecf1f3a9e894d0cc04
shows that the instance.info_cache.network_info is empty when we fail to
associated a floating IP with an instance.

This change adds more logging in the network layer around things that
would affect the instance_info_cache.network_info value, such as
building the network info model in nova-network and allocate/deallocate
of a fixed IP.

From the logs we have when this fails, get_instance_nw_info is never
returning anything for the instance that fails, which may mean that when
the fixed IP is allocated and associated to the instance, there is a
race where the fixed IP record in the database is disassociated
asynchronously while deallocating from the previous instance it was
associated with.

Related-Bug: #1249065

Change-Id: I42f9af1dabffdc98e69d9eeeab1875098bc71bc0