Updated openstack/openstack Project: openstack/nova  c746edfd05c1c0a068a5b64c26997606de9eb27b
Add virt resource update to ComputeNode object

The virt driver method get_available_resources() returns a dict
to the resource tracker. The fields of will be copied into a
ComputeNode object as an update. The copy is not exact as some
fields may be missing and some are named differently.

We add the method update_from_virt_driver() to the ComputeNode
to perform this update. It is not a remoteable method and does not
change the format or semantics of the object, so this change does
not bump the ComputeNode object version number.

This method will be used in a subsequent patch that replace the
compute_node dict in the resource tracker with a ComputeNode
object.

Change-Id: Ia010e428f400a5b4d0338e79670a0ef60e7d3b03