Updated openstack/openstack Project: openstack-infra/yaml2ical  81c242b7e0716cb1cd256b087b2ef06786603238
Update docs to indicate requirements and expand example meeting

Update the documentation to note which elements of the yaml
are required.  Also added the agenda_url and project_url keys
to the main doc list.

Lastly, updated the meeting/example.yaml to show usage of the new
duration, agenda_url, and project_url keys

Change-Id: Ifaa9f029ae920b22aa60ab3d47d14982707285d9