Updated openstack/openstack Project: openstack/neutron-lbaas  ea328e43a1b71eb0e8f72ae1ac1ee2f3ab03fe5a
Haproxy socket bind failing for IPv6 vip address

LBaaS is failing to spawn haproxy for ipv6 SLAAC subnet.
Haproxy socket binding to IPv6 VIP address is failing
as the IPv6 address is just created and is a tentative address.
So wait till the tentative flag is cleared and then
let haproxy is allowed to bind socket to this address.

Change-Id: I49a3c5fd4394eced7b67692ce28414a936ea107c
Closes-bug: #1403001