Updated openstack/openstack Project: openstack/cinder  5f6cbecc5ec5b6ca034290dc7bd04ab7d956cd15
Fix LIO target helper when missing targetcli

After patch to fix LIO target persistence was merged, a bug was
introduced where target creation/removal would result in an exception if
targetcli is not installed on the system.

This patch fixes this by checking if path to the default file exists on
the system and creating it if doesn't.

Change-Id: I5ecc1ae47482ea43a4890ccd81a4e872ab570069
Closes-Bug: #1453518