Updated openstack/openstack Project: openstack/ironic  2d5a6ca2c95c0789cc625379562cfa2122499ebb
Add option to configure passes in erase_devices

Today, there is no option to configure number of iterations to be
done for shred block device erasing and defaults it to 1. This patch
adds a configuration option to change the number of passes to be done
to erase a block device.

Change-Id: Ia870862f2abe830dee5e19a1c247e3f9399c6db2
Depends-On: I1921d33a6b364c4682b6c9baaf61ac092cfa11d7
Closes-Bug:#1465130