Updated openstack/openstack Project: openstack/nova  3443ffbae736f4fbbd5678dab0132021926a1099
cells: add devstack/tempest-dsvm-cells-rc for gating

We need to whittle down the large regex of excluded tests for the
tempest-dsvm-cells job but we also need to be able to test changes to
the regex, like removing entries that we think are now working.

This change moves the regex from project-config in the
tempest-dsvm-cells job definition to nova so that nova can control and
test the changes on it's own, since project-config is not currently able
to self-test job changes.

project-config change: I8cbc36266eadaf0dbef20e3e83f4874bcfc9ab3a

Change-Id: I2365619de4ba7b43a049771bac90d2deebaa3e9c