Updated openstack/openstack Project: openstack/oslo.versionedobjects  cf44278bc6cb037f974a8795b60cc5a600321582
Decouple ObjectVersionChecker and VersionedObjectRegistry

ObjectVersionChecker just needs a list of classes, we don't need
to check the base.VersionedObjectRegistry everytime, so just
initialize the list of classes in the constructor and default
it to the base.VersionedObjectRegistry to get the current
behavior. Will make it easy to run ObjectVersionChecker against
custom registries.

Change-Id: I700daa278473dff5285e0b21afffc03671e390f9