Updated openstack/openstack Project: openstack/tempest-lib  8f67d9f38ba3ef253c2e157767d2dcbe678cab10
Revert "Adding support for neutron-debug CLI."

There isn't a reason to have a function to call this command
live in tempest-lib. The neutron-debug CLI is only used for
local debug and isn't actually a client. It's for running
locally on the same node as neutron. It's completely out of
scope for living in tempest-lib. Just because there is a
CLI for something in an OpenStack project running doesn't
mean we need to have a dedicated function to call it. It's
a simple matter to locally subclass the CLIClient class and
add local functions to call commands for a particular test
suite. 

So let's revert this before it's included in a release. If
it were to be included in a release the effort to remove it
goes up quite a bit, because we need a proper deprecation
cycle.

This reverts commit 43c45b2b9dc6b5369556feb94431ed3e22185d34.

Change-Id: I8e77dd92abd482e6cdbaceff81dc12b9f224e2ec