Updated openstack/openstack Project: openstack-dev/devstack  ca4659931a9b108732a04fec0f3e66822d95807c
Don't install pbr in a venv

Because PIP_VIRTUAL_ENV was set for the installation of requirements,
and left around in scope, the installation of pbr no longer happened
in a global context, it instead landed inside the virtual
env. Unsetting the variable after requirements install gets us back to
where we expect.

This was an unintended side effect of the requirements-venv patch.

Change-Id: I2c4cb4305fec81a5fd237edabee78874ccd0da22