Updated openstack/openstack Project: openstack/nova  bc338f599d625e2d4a85a9774342bbf3781d58f8
Merge config_drive functional tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges config drive functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled for server tests to run
merged functional tests for both v2 and v2.1.

Change-Id: I17d3a057f1cf97241ec7c259c6ce943e7523f0d9