Updated openstack/openstack Project: openstack/governance  68927bd6da3144cc95bce6d989c9252844fe8bd5
add compute starter kit tag

This proposes a compute starter kit tag to specify a small number of
projects that are targeted at initial compute deploys by new
operators.

This is an evolution of the tag definition based on feedback provided
by folks that want to consume it, folks in favor, and folks
opposed. It makes for lengthy reading, but attempts to document the
questions that have been raised during the process.

Change-Id: Ie4aefe54a88aff9cebdf44e0e6c3f8ed95aa0d0e