Updated openstack/openstack Project: openstack/governance  331cc92f038432af63ad5391bb35bb848d70b94b
Ansible playbooks and roles to deploy OpenStack

Request to add the os-ansible-deployment project to OpenStack namespace.
os-ansible-deployment (OSAD) is a project dedicated to deploying OpenStack
clouds using Ansible and upstream OpenStack source code without the addition
of vendor driven packaging.

This project is already using all of the OpenStack infrastructure and services
and is gated through gerrit with testing being done on the whole stack using
tempest on our three supported branches.

We are committed to furthering our mission of a great
developer/deployer/operator experience and intend to build out the system to
support all of the OpenStack named services.

We use the OpenStack dev/operators mailing list (albeit infrequently) and hold
meetings twice weekly[1] (one in our channel and the other in an official
OpenStack meeting channel which is logged[2]).

The IRC channel #openstack-ansible is logged[3], and active with community
members, core developers and like minted people all helping and disusing issues
within the project as well as questions that come us from using Ansible to
operate and manage clouds.

The scope of os-ansible-deployment is to provide tooling for deploying OpenStack
using ansible which will include contributing to the efforts of upstream
Ansible. We are already working with the community to ensure that our efforts
are not duplicated or being made in direct contrast to what upstream Ansible is
already providing. While we do have some overlap within our currently carried
modules, the OpenStack client modules we have in tree, we are committed to moving
that functionality and code into upstream just as we've done with our other
plugins/modules[4][5].

With this project's move to "big tent" we'd only be looking to bring with us the
Master and Kilo branches as they've been designed to be community code with
all of the original Rackspace-isms removed. This would mean that the Juno and
Icehouse branches would be left behind in StackForge as they'd not be part of
the mainline project but we, the community of developers working on OSAD, will
be supporting those code bases until they're EOL'd by the upstream OpenStack
projects.

This commit is a formal request to add os-ansible-deployment into the
OpenStack namespace as an official project.

[1] https://wiki.openstack.org/wiki/Meetings/openstack-ansible
[2] http://eavesdrop.openstack.org/meetings/openstack_ansible_meeting/
[3] http://eavesdrop.openstack.org/irclogs/%23openstack-ansible/
[4] https://github.com/ansible/ansible/pull/9927
[5] https://github.com/ansible/ansible-modules-extras/pull/123

Change-Id: Id3f6175fcdd4a21019eabca56cd337145ff9c2b0