Updated openstack/openstack Project: openstack-infra/elastic-recheck  4a386668be83326069d7ad541903d5543e37380e
Update tox environment list

Remove unused tox environments and add used ones.

We don't gate on py26 or pypy support and are not planning on it any
time soon, so remove those environments from the default list.

Remove functional env since we don't use it and add queries environment
since we are now running it as non voting.

Change-Id: I573987ee49dbed7d520137044bbcef267e8761a7