Updated openstack/openstack Project: openstack-infra/nodepool  09b1ed53e4a4583731e19a96d4d246b02fc8628f
Trigger image builds using gearman

Move image building into its own module and use gearman to perform the
triggering of image builds. This is the first patch in a process to move
image building into a completely separate worker process.

Note that this changes the semantics of nodepool commands - they now do
not execute builds in the context of a new nodepool process but
communicate over gearman to an already running gearman worker.

Change-Id: Id0a64a8a957a5130f319ccd5beb3f45ea419eea6