Updated openstack/openstack Project: openstack/horizon  d6440886311d093396549747bf798df7719e2c56
Optimize getting unit in "metering" dashboard

Currently we using a meter-list for getting a "unit" for metric.
Meter list get all available metric on environment,
it's may be a thousands of records.
It's better to use sample-list for this metric with "limit" 1.
This request gets a 1 sample for queried metric.

Change-Id: I4bf0ab7ebcbcae1ddc90d8ee01d5f5a84ed4fbba
Related-bug: #1444456