Updated openstack/openstack Project: openstack/tempest  441d2cc917e67e0fa817ba6ab13e4018bbcd6873
Apply a naming rule of GET to compute clients(q*)

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to compute clients which names are "q*".

In addition, this patch removes unnecessary str() because
all callers specify string values and we need to pass py3
tests when migrating them into tempest-lib.

Partially implements blueprint consistent-service-method-names

Change-Id: Ifbb0a0bea4bfd1e94a054e1eafcc75e443ead1bb