Updated openstack/openstack Project: openstack/ceilometermiddleware  157097ef90b1fad48f8de90edb89010623148be5
add support to skip event/meter generation for certain projects

gnocchi is unuseable with swift backend because for every meter we
capture we are generating an event notification which creates 1+
meters, which creates 1+ event notifications, and so on. this will
flood the message queue and it will never catch up. by default, we
will configure system to ignore requests from gnocchi project.

DocImpact
Change-Id: I175c5d59529ebe51c8e5ccb379b54b05e9affe95