Updated openstack/openstack Project: openstack/heat  6bde4856c98b822fd1201ac44e35e04b81d2c9c0
Convergence: No back-up template for stack create

Current template need not be backed up while creating a stack. With
current logic, a circular reference is created with current stack
pointing to current template as previous template.

This happens when the template is stored before invoking stack create.
The template is stored along with the stack just before invoking the
converge_stack for creating the stack. The test cases are modified to
include instrutions to store the stack before running convergence.

Change-Id: I95b629d7d42201bc8273a3f70f3e3819211b32f8