Updated openstack/openstack Project: openstack/nova  1b26babb2fec776913e8b55b6b6577a9d87fe378
Merge security groups tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges security groups functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled to run
merged functional tests for both v2 and v2.1.

Change-Id: I45bc8a6ab4b48bff6ed28b4c779b575252161ed1