Updated openstack/openstack Project: openstack/trove-specs  fd2a2c1d579869d9f9d160179419c97c0879ae24
MySQL manager refactor

There are a number of forks of MySQL that differ to varying degrees.
We propose to create a class structure for MySQL-derived
datastores to avoid duplication of code dealing with features
and capabilities shared in common.

It is expected that the lessons learned from this effort will be
applicable to future efforts to provide differentiation for systems
such as MongoDB and PostgreSQL.

Change-Id: I56fe640a7f81f9a7a6906580e6484ddfbf84e677
Implements: blueprint mysql-manager-refactor