Updated openstack/openstack Project: openstack/requirements  8e04f4a252b0101512c0571667dca78b05972ba6
Remove swift from global-requirements.txt.

swift was added in 9b6a2b9050290bda548872cf444da6021f2aaeef as the
result of a mass script run over everything: we don't have any other
servers in global-requirements, and as swift doesn't offer a library
interface (and isn't meant to be on PyPI) it doesn't make sense to
have it.

Change-Id: Iba7dd6a705ed5bdc8836bcfd0f511f6bf0e00ed7