Updated openstack/openstack Project: openstack/oslo.versionedobjects  35138d5049dd0010a1475c5dfaf0df0a5ae8fe1d
Compound fields now coerce their elements' values

Added a special datastructures (CoercedList, CoercedDict and CoercedSet)
which are able to coerce their values to appropriate element type.
The instances of these structures are returned when the values of
appropriate compound fields initially coerced. As a result, all
subsequent calls which are going to add new elements to the fields'
colelctions will be processed by the overridden methods of new
datastructures, thus the new elements will be coerced and the exceptions
risen if the coercing is not possible.

This also fixes a bug in TestObjMakeList which was specifying incorrect
obj_name for the elements in 'objects' collection of MyList test class.
It was not leading to failures as the items of this collection were not
being coerced  prior to this patch.

Change-Id: I004646ce643e7e51305cfb1cd4eb5e8647a83513
Closes-Bug: #1466057