Updated openstack/openstack Project: openstack/oslo.versionedobjects  20e5ca83dc161c662bb024d6633dc250bcd8b77c
fields: allow subclasses in ObjectField/ListOfObjectsField

The Object class currently does a direct comparison of the
'obj_name()' value of the thing being coerced, against the
internal _obj_name field it was initialized with.

This does not allow for creation of object fields which
allow arbitrary subclasses - only the base class passes
the check.

Introduce a new 'subclasses' parameter, defaulting to
False, which traverses the python class hierarchy,
checking the obj_name() value on all parent classes
of the value being coerced.

Change-Id: Ie56b94e57a6e18292c10c926836592938540e515