Updated openstack/openstack Project: openstack/keystone  66dff30e84a60a0130add6e2a02131a993ed8101
Short names for auth plugins

A "default" entrypoint is defined for each auth method. The
default driver will be used if there's no config option for the
method, or the config option is not set, or if the config option
is set to "default".

For the external methods, since there's several of them, each gets
a short name that can be used rather than the qualified class.

bp stevedore

DocImpact

Change-Id: I2484af32e9eb3703869cf441e4f9851b54b0db2b