Updated openstack/openstack Project: openstack/tempest  3530c581dce7271c4e5255eab9a914226922fb1c
Ensure test_list_show_tenant_networks is isolated

This test can randomly fail when Neutron is enabled because
it assumes there is always at least a tenant (owned) network.

Unfortunately this is not true in Neutron because during concurrent
test runs other tests can create or delete shared/external networks
which are visible to all tenants.

This patch ensures that a network is always available.

Related-Bug: #1468523
Closes-Bug: #1491965

Change-Id: I90f8fce0733012b109e03bfe92755d0ba364bcae
Co-authored-by: Armando Migliaccio <armamig@gmail.com>