Updated openstack/openstack Project: openstack/glance  d2409ec7b367faa2e2d594a00b0001c40bda54f5
Provide extra parameter for FakeDB

In our codebase for tests we are often using same pattern to initialize
FakeDB instance which by default resets the database state and puts
sample data into it. After that we manually reset it's state, which
makes no sense. Here are the steps we are performing right now:

1. Reset database state
2. Puts data into database
3. Reset database state

It would be cleaner (and slightly faster) if we can skip these redundant
steps (2,3) in case we need an empty database. Provided fix allows
developers who are writting tests to decide if sample data should be
initialized or not.

Change-Id: I9e03b1c26d9ad311b8ff4330d19b0db93e6d1157