Updated openstack/openstack Project: openstack-infra/project-config  a2964c0ce25185cb6e5374d186e0d7d556b04bf8
Add temporary non voting multinode smoke jobs

As a first step in making grenade multinode, we need to figure out how
reliable tempest smoke jobs are in multinode.  As it is tempest-full
fails *roughly* 10% of the time with nova-network, and neutron (with
DVR) fails every time. But we are turning off DVR for this job, as we
don't expect grenade to do DVR just yet, so I don't have any numbers on how
reliable tempest-full is with non-dvr neutron.

Running these jobs as non-voting everywhere the full jobs were run and
then some in order to get a decent dataset in logstash.openstack.org and
graphite.

We cannot proceed with multinode grenade until these jobs are stable.
Once grenade is multinode, these jobs can be removed.

Change-Id: I5bb261787f59650b71dbdf75bdd787bd2efb4f9f