Updated openstack/openstack Project: openstack/neutron  b257136e1c68291f07ae47b8592dc407a7a40cd9
Make DHCPv6 out of bounds API test deterministic

The test_dhcp_stateful_fixedips_outrange API test was randomly
picking an IP from last + 1 up to last + 10 in the allocation
range. This made it fail randomly when there was an issue related
to the subnet allocation having an off-by-one issue.

This adjusts the test to just always test last +1 and +2.

Related-Bug: #1468163
Change-Id: I641ab092e0ea0aae67ec717b492118a2f8a6f4fd