Updated openstack/openstack Project: openstack/nova  811fc4e27f167e6d4f6c1a9c9269190f09b11f17
Let soft-deleted instance_system_metadata readable

Currently, We can't delete instance system metadata after instance
is deleted because of following reason:
1) We might have 'changed-since' request which will access
the soft deleted system_metadata
2) We might has period audit function which also access the
soft deleted system_metadata

The root cause is the join function filtered the 'deleted'
data of system metadata and it lead to ValueError later.

Conceptually the system metadata is associated with instance,
it will be used in instance's lifecycle, the data will only
be updated so we won't have 2 copies or more system_metadata
for the same instance.
So allow to read it even if it's deleted is safe.

This patch load the system_metadata from db when instance is
in deleted state to avoid can't delete system_metadata issue.

Co-Authored-By: Alex Xu <hejie.xu@intel.com>

Change-Id: Ifd8d2c91879834d3759aafdb036dac209a246d1b
Partial-Bug: #1226049