Updated openstack/openstack Project: openstack/nova  6f3448bdbff15c268cc5b5f9f0785523b09c6507
Use oslo-config-generator instead of generate_sample.sh

Oslo team is actively getting rid of the old style generator.py
in openstack/common/config. So we should switch to the newer
oslo-config-generator which is the way forward. For simplicity's
sake, i used one single list_opts in opts.py. However we can
split this up and add entrypoints in setup.cfg if we decide to
do that instead.

blueprint oslo-config-generator

Change-Id: Ib7c72af7be47df08983708ac6da0ce103518b78c