Updated openstack/openstack Project: openstack/nova  1e07fe6d003e1fa96e1b51cfa3b63e1c7e23196a
Cells: Skip initial sync of block_device_mapping

When initially creating the block device mapping in a child cell it is
synced up to the api without a device name. Since later syncs rely on
the device name to find and update the block device mapping versus
creating a new one we should skip this sync and wait for the later
update.

Change-Id: I683580f668158e8dea5714105946535ee6124e51
Closes-Bug: 1444128