Updated openstack/openstack Project: openstack/nova  edd5ad55a96c06e71542fb2582ff190baaf8492b
hypervisor support matrix: add feature "serial console"

The "serial console" feature was introduced during the Juno cycle
and is an alternative to remote access via VNC, RDP, SPICE for
platforms which don't support these.
This change adds a new feature row for that to the hypervisor support
matrix table. Because I'm not able to make a statement for all
hypervisor targets I use the "unknown" state for them.

Change-Id: I23381e30b6ddb823a50f7755179eb02bc18222e8