Updated openstack/openstack Project: openstack-infra/shade  354683a12e191f1f9710643858dd85e82c94b4fe
Fix available_floating_ip when using Nova network

When neutron is not available and there are no floating IP already
allocated to the project, available_floating_ip calls meta.obj_to_dict()
on a Bunch object.

Moreover we can save an API call as we already know the pool name.

Change-Id: I997cb0edc4fd9160b6804a29d404190526303720