Updated openstack/openstack Project: openstack/neutron-fwaas  2e3d4ac83d4616328336f8f7299f3a6302663c39
Fixes firewall going to error state on an update

Whenever an attribute of firewall rule, firewall policy or the firewall
is updated, the firewall state is set to 'error'. The current code in the
update_firewall method handles only the cases where routers are either
added, removed or both from a firewall. This commit also handles setting
the status correctly when other attributes are updated.

Co-Authored by: vikram.choudhary <vikram.choudhary@huawei.com>
Closes-Bug: #1455863

Change-Id: I32757cbbdaaea0547f653bd9bbdb9ccd2a0780ac