Updated openstack/openstack Project: openstack/nova  227fcadb649af74422eb89225f316280429311c6
Raise invalid input if use invalid ip for network to attach interface

When posting an 'attach interface' request to Nova with network id and a fixed
ip which does not belong to this network, Nova needs to raise invalid input
exception, instead of HTTP 500, in order to inform the user to correct the
request. This code change will take effect with Neutron only.

Change-Id: Ie08858278cb9e819b7f434567d886367b20f09f1
Closes-Bug: 1369871