Updated openstack/openstack Project: openstack/nova  d3c6e641f9662d02d4ba6a96caa7e0067f6d75d9
hypervisor support matrix: fix "evacuate" for s390 and hyper-v

Misled by the title of the feature "evacuate" the platform s390x
marked it as "complete". After a discussion in Nova IRC channel it was
clarified that this should be called "set host maintenance". This is
"missing" for s390x just like on x86. hyper-v doesn't implement the
necessary methods too.

http://lists.openstack.org/pipermail/openstack-dev/2015-June/067787.html

Change-Id: I3cd386269868c9fbbdbb16ca6281a86f4e2f1e0f