Updated openstack/openstack Project: openstack/neutron  462c72daaa874c54aba3d51a3559845145651f00
Add deadlock retry to API and ML2 RPC port update

With the switch to the pymsql SQL driver, eventlet will now yield
during database transactions. This greatly increased our probability
of multiple coroutines running transactions on the same table that
result in deadlocks.

These deadlocks could result from many things including the following:
* a coroutine holding a pessimistic "SELECT for UPDATE" lock when
  another tries to update the locked records
* two coroutines both issue update statements using a WHERE clause
  invalidated by the other update (e.g. from a compare and swap approach)
* two coroutines insert records that, when combined, violate a unique
  constraint on the table in a master-master Galera deployment
* any two workers using "SELECT for UPDATE" in a master-master Galera
  deployment (write-set certification failure translates to deadlock)

This problem is exacerbated by the switch to multiple API and RPC
workers, each of which can lead to most of the errors above even
without the switch to pymysql.

This patch adds a deadlock retry decorator to the delete, create,
and update methods at the HTTP API layer. Additionally, it adds a
decorator to the update_port_status AMQP API in ML2 since it updates
the port table, which is a heavily locked table by ML2 making it a
prime candidate for deadlocks.

Nova has had relied on the deadlock retry mechanism for quite some
time now. We were limping along by not using additional workers and
by relying on the unyielding nature of the MySQL C driver to
serialize everything.

Closes-Bug: #1464612
Change-Id: I635cc49ca69f589f99ab145d4d51e511b24194d2