Updated openstack/openstack Project: openstack/diskimage-builder  ba6959bb92e667c48474929390df2924d4c03ce5
Allow source-repositories ref to be "*"

Passing a source-repositories ref of "*" should signal fetching all
heads similar to when a non-cached ref is requested. Reuse the same
fallback logic, but skip unnecessary checks since "*" is not a real
refname. Also expand the fallback to update tags, and to --purge
local refs that no longer exist on the remote for additional safety.

Change-Id: I4562c9689a8d235ebe09b2f7178aa5890dbc85f1